+++ CHANGELOG.md
@@ -8,9 +8,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 
-- *CUMULUS-2895**
+- **CUMULUS-2895**
   - Updated KMS code to aws sdk v3
-- *CUMULUS-2899**
+- **CUMULUS-2888**
+  - Update CloudWatch Events code to AWS SDK v3
 - **CUMULUS-2893**
   - Updated Kinesis code to AWS SDK v3